src/main/java/com/woowacourse/zzinbros/user/service/SearchService.java
package com.woowacourse.zzinbros.user.service;

import com.woowacourse.zzinbros.user.domain.User;
import com.woowacourse.zzinbros.user.domain.repository.UserRepository;
import com.woowacourse.zzinbros.user.dto.UserResponseDto;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.Set;
import java.util.stream.Collectors;

@Transactional
@Service
public class SearchService {
    private final UserRepository userRepository;

    public SearchService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Transactional(readOnly = true)
    public Set<UserResponseDto> search(String searchName, Pageable pageable) {
        List<User> users = userRepository.findByNameNameContaining(searchName, pageable);
        return users.stream()
                .map(UserResponseDto::new)
                .collect(Collectors.toSet());
    }
}
